pyproject.toml
[project]
name = "python-cmdb"
version = "0.1.0"
description = "FastAPI based CMDB"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic",
    "bcrypt",
    "boto3",
    "fastapi[all]",
    "loguru",
    "pydantic==2.5.3",
    "pymysql",
    "redis",
    "sqlalchemy",
    "uvicorn",
    "fastapi-users[sqlalchemy]",
    "fastapi-users-db-sqlalchemy",
    "python-jose[cryptography]",
    "passlib[argon2,bcrypt]",
    "python-multipart",
    "setuptools",
    "aiomysql",
    "fastapi-authz",
    "casbin",
    "casbin-sqlalchemy-adapter>=1.4.0",
]
